DCWL-2003-v2 moved services 0.5
