+++ app/uk/gov/hmrc/customs/declarations/information/services/declarations/DeclarationFullService.scala
@@ -20,7 +20,8 @@ import uk.gov.hmrc.customs.api.common.controllers.ErrorResponse
 import uk.gov.hmrc.customs.api.common.controllers.ErrorResponse.ErrorInternalServerError
 import uk.gov.hmrc.customs.declarations.information.connectors.{ApiSubscriptionFieldsConnector, DeclarationFullConnector}
 import uk.gov.hmrc.customs.declarations.information.logging.InformationLogger
-import uk.gov.hmrc.customs.declarations.information.services.{FullResponseFilterService, InformationConfigService, UniqueIdsService}
+import uk.gov.hmrc.customs.declarations.information.services.filters.FullResponseFilterService
+import uk.gov.hmrc.customs.declarations.information.services.{InformationConfigService, UniqueIdsService}
 import uk.gov.hmrc.http.HttpResponse
 
 import javax.inject.{Inject, Singleton}
