+++ app/uk/gov/hmrc/customs/declarations/information/services/declarations/DeclarationSearchService.scala
@@ -21,7 +21,8 @@ import uk.gov.hmrc.customs.api.common.controllers.ErrorResponse
 import uk.gov.hmrc.customs.api.common.controllers.ErrorResponse.ErrorInternalServerError
 import uk.gov.hmrc.customs.declarations.information.connectors.{ApiSubscriptionFieldsConnector, DeclarationSearchConnector}
 import uk.gov.hmrc.customs.declarations.information.logging.InformationLogger
-import uk.gov.hmrc.customs.declarations.information.services.{InformationConfigService, SearchResponseFilterService, UniqueIdsService}
+import uk.gov.hmrc.customs.declarations.information.services.filters.SearchResponseFilterService
+import uk.gov.hmrc.customs.declarations.information.services.{InformationConfigService, UniqueIdsService}
 import uk.gov.hmrc.http.HttpResponse
 
 import javax.inject.{Inject, Singleton}
