+++ app/uk/gov/hmrc/customs/declarations/information/services/filters/FullResponseFilterService.scala
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package uk.gov.hmrc.customs.declarations.information.services
+package uk.gov.hmrc.customs.declarations.information.services.filters
 
 import uk.gov.hmrc.customs.declarations.information.xml.HelperXMLUtils.extractNamespacesFromAllElements
