+++ test/unit/services/DeclarationFullResponseFilterServiceSpec.scala
@@ -19,7 +19,6 @@ package unit.services
 import org.scalatest.Assertion
 import play.api.test.Helpers
 import uk.gov.hmrc.customs.api.common.xml.ValidateXmlAgainstSchema
-import uk.gov.hmrc.customs.declarations.information.services.FullResponseFilterService
 import util.FullTestXMLData.backendDeclarationFullResponse
 import util.UnitSpec
 
@@ -29,6 +28,7 @@ import scala.util.Try
 import scala.xml._
 import scala.xml.transform.{RewriteRule, RuleTransformer}
 import ValidateXmlAgainstSchema._
+import uk.gov.hmrc.customs.declarations.information.services.filters.FullResponseFilterService
 
 class DeclarationFullResponseFilterServiceSpec extends UnitSpec {
   implicit val ec: ExecutionContext = Helpers.stubControllerComponents().executionContext
